files splitting and renames
