+++ src/corona_hakab_model/common/medical_state.py
@@ -1,7 +1,7 @@
 from abc import ABC
 
-from common.state_machine import State  # , StochasticState, TerminalState
-from common.util import BucketDict
+from .state_machine import State  # , StochasticState, TerminalState
+from .util import BucketDict
 
 
 class MedicalState(State, ABC):
