+++ src/corona_hakab_model/generation/geographic_circle.py
@@ -1,7 +1,8 @@
 from typing import List
 
 import numpy as np
-from common.circles import Circle, SocialCircle
+from common.circle import Circle
+from common.social_circle import SocialCircle
 from generation.circles_consts import GeographicalCircleDataHolder
 from generation.connection_types import ConnectionTypes, In_Zone_types, Multi_Zone_types, Education_Types
