+++ src/corona_hakab_model/supervisor.py
@@ -1,24 +1,20 @@
 # flake8: noqa flake8 doesn't support named expressions := so for now we have to exclude this file for now:(
-
 from __future__ import annotations
-
 from abc import ABC, abstractmethod
 from functools import lru_cache
 from typing import Any, Callable, List, NamedTuple, Sequence, Union, Dict
-
 from pathlib import Path
 import manager
-from common.state_machine import StochasticState
-
 import numpy as np
 import pandas as pd
 
+from common.state_machine import StochasticState
 from typing import TYPE_CHECKING
+from common.histogram import TimeHistograms
 
 if TYPE_CHECKING:
     from common.state_machine import State
 
-from common.histogram import TimeHistograms
 
 class SimulationProgression:
     """
